pyproject.toml
[tool.poetry]
name = "ndsspeech"
version = "0.1.0"
description = ""
authors = ["mmcux <marcel.meyer@uni-paderborn.de>", "ypix <ypix@w47.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
transformers = "^4.41.2"
numpy = "<2.0.0"
python-dotenv = "^1.0.1"
poethepoet = "^0.27.0"
torch = { version = "^2.3.1+cu121", source = "pytorch-gpu" }
tts = "^0.22.0"
tensorboard = "^2.17.0"


[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
